some bug fixes

+++ src/api/v1/engines/user_inputs.py
@@ -14,7 +14,7 @@ from src.core.enums import (
 
 class LoraIDAndWeightInput(BaseModel):
     lora_model_id: int
-    weight: int
+    weight: float
 
 
 class EngineUserInput(BaseModel):
